feat: add queue specifier option
ref: https://app.asana.com/0/1198103229139841/1202490157157977

This change adds the option `kill_queues` to specify which queues
we want the sidekiq-worker-killer to actually request shutdown on.
By default the option is empty and the worker-killer will act on
all queues; if a non-empty array is passed into `kill_queues` at
initialization, then the worker-killer will only act on the queues
within said array and no other queues.